SpermAnalyzerAI/model/requirements.txt
# SpermAnalyzerAI - Model Training Requirements
# Basic deep learning and computer vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data processing and augmentation
albumentations>=1.3.0
pandas>=2.0.0
Pillow>=10.0.0
imageio>=2.31.0

# Machine learning optimization
onnx>=1.14.0
onnxruntime>=1.15.0
tensorrt>=8.6.0; platform_machine=="x86_64" and platform_system=="Linux"

# Visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0
psutil>=5.9.0

# For data science
jupyter>=1.0.0
ipywidgets>=8.0.0